sportsbet-tennis/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
duckdb>=0.10
SQLAlchemy>=2.0
pydantic>=2.5
fastapi>=0.110
uvicorn>=0.29
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
joblib>=1.3
matplotlib>=3.8
streamlit>=1.34
typer>=0.12

lightgbm>=3.3
